@@ -18,13 +18,15 @@ android {
             setV2SigningEnabled(true)
         }
     }
-    compileSdkVersion 24
+    buildToolsVersion "28.0.3"
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.b1b.js.erpandroid_kf"
         minSdkVersion 16
-        targetSdkVersion 22
-        versionCode 59
-        versionName "6.7"
+        targetSdkVersion 26
+        versionCode 63
+        versionName "7.1"
+        vectorDrawables.useSupportLibrary = true
         ndk {
             abiFilters "armeabi", "armeabi-v7a"
         }
