@@ -1,10 +1,9 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 24
+    compileSdkVersion 28
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 22
         versionCode 10
         versionName "3.5"
     }
@@ -22,6 +21,7 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:support-v4:24.2.1'
-    implementation 'com.android.support:appcompat-v7:24.2.1'
+//    implementation 'com.android.support:support-v4:24.2.1'
+//    implementation 'com.android.support:appcompat-v7:24.2.1'
+//    implementation 'com.android.support:appcompat-v7:28.0.0'
 }
