@@ -1,7 +1,6 @@
 package com.b1b.js.erpandroid_kf;
 
 import android.annotation.SuppressLint;
-import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.os.Bundle;
@@ -10,15 +9,16 @@ import android.os.Looper;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
+import android.support.v4.app.FragmentPagerAdapter;
 import android.support.v4.view.PagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
+import android.util.DisplayMetrics;
 import android.util.Log;
-import android.util.LruCache;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.ViewParent;
 import android.widget.ImageView;
 import android.widget.TextView;
 
@@ -28,10 +28,13 @@ import com.b1b.js.erpandroid_kf.task.TaskManager;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.LinkedList;
 import java.util.List;
 
-import utils.net.HttpUtils;
 import utils.common.MyImageUtls;
+import utils.framwork.BitmapLruCache;
+import utils.net.HttpUtils;
 
 public class PicDetailActivity extends AppCompatActivity {
 
@@ -40,6 +43,8 @@ public class PicDetailActivity extends AppCompatActivity {
     private List<ZoomImageView> mImgs;
     private List<String> paths;
     private PagerAdapter adapter;
+    public static final String ex_Path = "path";
+    public static final String ex_Paths = "paths";
     int pos;
     TextView tv;
     @Override
@@ -49,26 +54,10 @@ public class PicDetailActivity extends AppCompatActivity {
         zoomIv = (ZoomImageView) findViewById(R.id.activity_pic_detail_iv);
         mViewPager = (ViewPager) findViewById(R.id.activity_pic_detail_viewpager);
         tv = (TextView) findViewById(R.id.activity_pic_detail_tv);
-        String path = getIntent().getStringExtra("path");
-        paths = getIntent().getStringArrayListExtra("paths");
-        mImgs = new ArrayList<>();
-        ZoomImageView zoomImageView = new ZoomImageView(PicDetailActivity.this);
-        try {
-            Bitmap bitmap = MyImageUtls.getMySmallBitmap(path, 500, 500);
-            zoomImageView.setImageBitmap(bitmap);
-            zoomImageView.setTag(bitmap);
-            mImgs.add(zoomImageView);
-        } catch (OutOfMemoryError error) {
-            error.printStackTrace();
-        }
+        String path = getIntent().getStringExtra(ex_Path);
+        paths = getPaths();
         pos = getIntent().getIntExtra("pos", 0);
-        FragmentManager mgr = getSupportFragmentManager();
-//        paths = new ArrayList<>();
-//        paths.add("https://ss2.baidu.com/6ONYsjip0QIZ8tyhnq/it/u=1754781758,2069282087&fm=173&app=49&f=JPEG?w=612&h=365&s=5AD2688C42520FD44484788A0300E09C");
-//        paths.add("https://ss0.baidu.com/6ONWsjip0QIZ8tyhnq/it/u=2769768633,2874225384&fm=173&app=49&f=JPEG?w=640&h=451&s=FEC7A144062BB7551C5529070300A0C2");
-//        paths.add("https://ss2.baidu.com/6ONYsjip0QIZ8tyhnq/it/u=714297440,3631358855&fm=173&app=49&f=JPEG?w=640&h=620&s=082193194C1372DC0839D5D20300D0A1");
-//        paths.add("https://ss0.baidu.com/6ONWsjip0QIZ8tyhnq/it/u=1790721534,1107844676&fm=173&app=49&f=JPEG?w=640&h=821&s=6F32038D5402FEF984000CF60300D02F");
-        adapter = new ImgFragmentAdapter(mgr, paths);
+        adapter = getPagerAdapter(paths);
         mViewPager.setAdapter(adapter);
         mViewPager.addOnPageChangeListener(new ViewPager.OnPageChangeListener() {
             @Override
@@ -90,93 +79,65 @@ public class PicDetailActivity extends AppCompatActivity {
         mViewPager.setCurrentItem(pos);
     }
 
+    public List<String> getPaths() {
+        List<String> paths = getIntent().getStringArrayListExtra(ex_Paths);
+        if (paths == null) {
+            paths = new ArrayList<>();
+        }
+        return paths;
+    }
+
+    public PagerAdapter getPagerAdapter(List<String> paths) {
+        FragmentManager mgr = getSupportFragmentManager();
+        PagerAdapter mAdapter = new ImgFragmentAdapter(mgr, paths);
+        return mAdapter;
+    }
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        ((ImgFragmentAdapter) adapter).releasBitmap();
-    }
-
-    class PicDetailViewPagerAdapter extends PagerAdapter {
-        public PicDetailViewPagerAdapter(Context mContext, List<ZoomImageView> data) {
-            this.mContext = mContext;
-            this.data = data;
-        }
-
-        private Context mContext;
-        private List<ZoomImageView> data;
-
-        @Override
-
-        public int getCount() {
-            return data == null ? 0 : data.size();
-        }
-
-        @Override
-        public boolean isViewFromObject(View view, Object object) {
-            return view == object;
-        }
-
-        @Override
-        public Object instantiateItem(ViewGroup container, int position) {
-            ZoomImageView zoomImageView = data.get(position);
-            ViewGroup parent = (ViewGroup) zoomImageView.getParent();
-            if (parent != null) {
-                parent.removeView(zoomImageView);
-            }
-            container.addView(zoomImageView);
-            return zoomImageView;
-        }
-
-        @Override
-        public void destroyItem(ViewGroup container, int position, Object object) {
-            View zoomVIew = (View) object;
-            Bitmap tag = (Bitmap) zoomVIew.getTag();
-            container.removeView((View) object);
-        }
-
-        @Override
-        public int getItemPosition(Object object) {
-            return PagerAdapter.POSITION_NONE;
+        try {
+            ((ImgFragmentAdapter) adapter).releasBitmap();
+        } catch (Exception e) {
+            e.printStackTrace();
         }
     }
 
-    static class ImgFragmentAdapter extends FragmentStatePagerAdapter {
-        private List<String> imgs = new ArrayList<>();
-
-
-        private LruCache<String, Bitmap> cacheMap;
-        public ImgFragmentAdapter(FragmentManager fm) {
-            super(fm);
-        }
+    static class ImgFragmentAdapter extends FragmentPagerAdapter {
+        private List<String> imgs;
+        private LinkedHashMap<String, ImgFragment> mFrags = new LinkedHashMap<>();
+        private BitmapLruCache<String> cacheMap;
+        FragmentManager mFragmentManagerfMgr;
 
         public ImgFragmentAdapter(FragmentManager fm, List<String> imgs) {
             super(fm);
+            this.mFragmentManagerfMgr = fm;
             this.imgs = imgs;
-
-            int maxMemory = (int) Runtime.getRuntime().maxMemory();
-            Log.e("zjy", "PicDetailActivity->ImgFragmentAdapter(): MaxMem==" + maxMemory);
+            int totalSize = (int) Runtime.getRuntime().maxMemory();
+            int maxMemory = (int) Runtime.getRuntime().freeMemory();
+            float rate = 1024 * 1024;
+//            Log.e("zjy", "PicDetailActivity->ImgFragmentAdapter(): MaxMem==" + totalSize / rate +
+//                    "MB");
+//            Log.e("zjy",
+//                    "PicDetailActivity->ImgFragmentAdapter(): freeSize==" + maxMemory / rate + "MB");
             // 取处内存的 1/5 用来当 缓存 大小
-            int cachSize = maxMemory / 3;
-            Log.e("zjy", "PicDetailActivity->ImgFragmentAdapter(): cacheSize==" + cachSize);
+            int cachSize = maxMemory / 2;
+            Log.e("zjy", "PicDetailActivity->ImgFragmentAdapter(): cacheSize==" + cachSize / rate);
             // 实例化 LruCache
-            cacheMap=new LruCache<String, Bitmap>(cachSize) {
-                //内部方法sizeOf设置每一张图片的缓存大小
-                protected int sizeOf(String key, Bitmap value) {
-                    //在每次存入缓存时调用，告诉系统这张缓存图片有多大
-                    // 相当于 为每次 要缓存的 资源 分配 大小空间
-                    if (value == null) {
-                        MyApp.myLogger.writeBug("value is Null,key=" + key + "\t to ");
-                        return 0;
-                    }
-                    return value.getByteCount();
-                }
-            };
+            cacheMap = new BitmapLruCache<>(cachSize);
         }
+
 
         @Override
         public Fragment getItem(int position) {
-            String s = imgs.get(position);
-            Fragment picFrag = ImgFragment.newInstance(s, cacheMap);
+            String imgUrl = imgs.get(position);
+            ImgFragment picFrag = mFrags.get(imgUrl);
+
+            if (picFrag == null) {
+                Log.e("zjy", getClass() + "->getItem():new Instatce ==" + imgUrl);
+                picFrag = ImgFragment.newInstance(imgUrl, cacheMap);
+                mFrags.put(imgUrl, picFrag);
+            }
+            Log.e("zjy", getClass() + "->getItem(): ==" + picFrag.toString());
             return picFrag;
         }
 
@@ -185,7 +146,7 @@ public class PicDetailActivity extends AppCompatActivity {
             return imgs.size();
         }
 
-        public void releasBitmap(){
+        public void releasBitmap() {
             cacheMap.evictAll();
         }
     }
@@ -193,135 +154,178 @@ public class PicDetailActivity extends AppCompatActivity {
     public static class ImgFragment extends Fragment {
         private Handler mHandler = new Handler(Looper.getMainLooper());
         String imgPath = "";
+        private View itemView;
         ImageView mImgView;
-        Bitmap nBmp =null;
-        private LruCache<String, Bitmap> cacheMap;
-        public ImgFragment(){
+        Bitmap nBmp = null;
+        private BitmapLruCache<String> cacheMap;
+
+        public ImgFragment() {
 
         }
+
+        public void setImgPath(String imgPath) {
+            this.imgPath = imgPath;
+            if (imgPath != null) {
+                int i = imgPath.lastIndexOf("/");
+                if (i >= 0) {
+                    tag = imgPath.substring(imgPath
+                            .lastIndexOf("/") + 1);
+                }
+            }
+            tag = toString() + "/" + tag;
+        }
+
+        LinkedList<ImgFragment> mFragCache;
+
+        @SuppressLint("ValidFragment")
+        private ImgFragment(BitmapLruCache<String> cacheMap, LinkedList<ImgFragment> mFragCache) {
+            this.cacheMap = cacheMap;
+            this.mFragCache = mFragCache;
+        }
+
+        int heightPixels = 500;
+        int widthPixels = 100;
+        private String tag = "未知";
+
         @SuppressLint("ValidFragment")
         public ImgFragment(String imgPath) {
             this.imgPath = imgPath;
         }
+
         @SuppressLint("ValidFragment")
-        public ImgFragment(String imgPath, LruCache<String, Bitmap> cacheMap) {
+        public ImgFragment(String imgPath, BitmapLruCache<String> cacheMap) {
             this.imgPath = imgPath;
             this.cacheMap = cacheMap;
+            if (imgPath != null) {
+                int i = imgPath.lastIndexOf("/");
+                if (i >= 0) {
+                    tag = imgPath.substring(imgPath
+                            .lastIndexOf("/") + 1);
+                }
+            }
+            tag = toString() + "/" + tag;
         }
 
         public static ImgFragment newInstance(String path) {
             return new ImgFragment(path);
         }
 
-        public static ImgFragment newInstance(String path, LruCache<String, Bitmap> cacheMap) {
+        public static ImgFragment newInstance(String path, BitmapLruCache<String> cacheMap) {
             return new ImgFragment(path, cacheMap);
         }
+
+        public static ImgFragment newInstance(BitmapLruCache<String> cacheMap, LinkedList<ImgFragment> mFragCache) {
+            return new ImgFragment(cacheMap, mFragCache);
+        }
+
         @Nullable
         @Override
         public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle
                 savedInstanceState) {
-            Bundle arguments = getArguments();
-            View itemView = inflater.inflate(R.layout.frag_pic_detail, container, false);
+            Log.e("zjy", "->onCreateView(): ==" + tag);
+            //            itemView = getView();
+            if (itemView == null) {
+                itemView = inflater.inflate(R.layout.frag_pic_detail, container, false);
+            } else {
+                Log.e("zjy", "initView->onCreateView():useCacheView ==" + tag);
+            }
+            ViewParent parent = itemView.getParent();
+            if (parent != null) {
+                ViewGroup view = (ViewGroup) parent;
+                view.removeView(itemView);
+                Log.e("zjy", "initView->onCreateView(): ==hasView" + tag);
+            } else {
+                Log.e("zjy", "initView->onCreateView():noParent ==" + tag);
+            }
             mImgView = (ImageView) itemView.findViewById(R.id.frag_pic_detail_iv);
-            Bitmap tempBmp = cacheMap.get(imgPath);
-            if (tempBmp != null) {
-               // nBmp = tempBmp;
-                Log.e("zjy", "PicDetailActivity->onCreateView(): getCacheBitmap==" + tempBmp);
-                mImgView.setImageBitmap(tempBmp);
-            }else{
-                Runnable imgRun = new Runnable() {
-                @Override
-                public void run() {
-                        Log.e("zjy", "PicDetailActivity->run(): Decode==" + imgPath.substring(imgPath
-                                .lastIndexOf("/") + 1));
-                    final Bitmap decodeBmp = loadImg(imgPath);
-                    if (decodeBmp != null) {
-                        cacheMap.put(imgPath, decodeBmp);
-                        mHandler.post(new Runnable() {
-                            @Override
-                            public void run() {
-                                mImgView.setImageBitmap(decodeBmp);
+            DisplayMetrics displayMetrics = getActivity().getResources().getDisplayMetrics();
+            heightPixels = displayMetrics.heightPixels;
+            widthPixels = displayMetrics.widthPixels;
+            if (imgPath == null) {
+                mImgView.setImageBitmap(null);
+            } else {
+                Bitmap tempBmp = cacheMap.get(imgPath);
+                if (tempBmp != null) {
+                    // nBmp = tempBmp;
+                    Log.e("zjy", "PicDetailActivity->onCreateView(): getCacheBitmap==" + tempBmp);
+                    mImgView.setImageBitmap(tempBmp);
+                } else {
+                    Runnable imgRun = new Runnable() {
+                        @Override
+                        public void run() {
+                            final Bitmap decodeBmp = loadImg(imgPath);
+                            if (decodeBmp != null) {
+                                if (decodeBmp.isRecycled()) {
+                                    Log.e("zjy", "->run(): recycle img==" + imgPath);
+                                    return;
+                                }
+                                cacheMap.put(imgPath, decodeBmp);
+                                mHandler.post(new Runnable() {
+                                    @Override
+                                    public void run() {
+                                        mImgView.setImageBitmap(decodeBmp);
+                                    }
+                                });
+                            } else {
+                                Log.e("zjy", "-load bitmap Failed ,key=" + tag);
+                                MyApp.myLogger.writeBug("load bitmap Failed ,key=" + imgPath);
                             }
-                        });
-                        Log.e("zjy", "PicDetailActivity->run(): BitampSize==" + decodeBmp.getByteCount());
-                    }else{
-                        MyApp.myLogger.writeBug("load bitmap Failed ,key=" + imgPath);
-                    }
+                        }
+                    };
+                    TaskManager.getInstance().execute(imgRun);
                 }
-            };
-            TaskManager.getInstance().execute(imgRun);
             }
             return itemView;
         }
 
-        /*@Override
-        public void onStop() {
-            super.onStop();
-            mImgView.setImageBitmap(null);
-            Log.e("zjy", "PicDetailActivity->onStop(): --" + imgPath.substring(imgPath.lastIndexOf("/") + 1)
-                    + toString());
-        }
-*/
-        @Override
-        public void onDestroyView() {
-            super.onDestroyView();
-            Log.e("zjy", "PicDetailActivity->onDestroyView(): --" + imgPath.substring(imgPath.lastIndexOf("/")+1)
-                    + toString());
-            mImgView.setImageBitmap(null);
-            Bitmap remove = cacheMap.remove(imgPath);
-            if (remove != null) {
-                remove.recycle();
-            }
-        }
-
-        Bitmap loadImg(String path) {
-            if (path.startsWith("http")||path.startsWith("https")) {
-                try {
-                    InputStream inputStream = HttpUtils.create(path).getInputStream();
-                    return BitmapFactory.decodeStream(inputStream);
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-                return null;
-            }else {
-                try {
-                    return BitmapFactory.decodeFile(imgPath);
-                } catch (Throwable e) {
-                    e.printStackTrace();
-                }
-                return null;
-            }
-        }
-        @Override
-        public void onDestroy() {
-            super.onDestroy();
-            try {
-                mImgView.setImageBitmap(null);
-            } catch (Throwable e) {
-                e.printStackTrace();
-            }
-
-            Log.e("zjy", "PicDetailActivity->onDestroy(): --" + imgPath.substring(imgPath.lastIndexOf("/")+1)
-                    + toString());
-        }
-
-   /*
-
-        @Override
-        public void onAttach(Context context) {
-            Log.e("zjy", "PicDetailActivity->onAttach(): ++" + imgPath.substring(imgPath.lastIndexOf("/")+1)
-                    + toString());
-            super.onAttach(context);
-        }
 
         @Override
         public void onDetach() {
             super.onDetach();
-            Log.e("zjy", "PicDetailActivity->onDetach(): --" + imgPath.substring(imgPath.lastIndexOf("/")+1)+ toString());
-           *//* if (nBmp != null) {
-                nBmp.recycle();
-                nBmp = null;
-            }*//*
-        }*/
+            if (mFragCache != null) {
+                mFragCache.offer(this);
+            }
+            Log.e("zjy", "->onDetach(): ==" + tag);
+        }
+
+        @Override
+        public void onPause() {
+            super.onPause();
+            itemView = getView();
+            Log.e("zjy", "->onPause(): ==" + tag);
+        }
+
+
+        Bitmap loadImg(String path) {
+            if (path.startsWith("http") || path.startsWith("https")) {
+                Bitmap mBitmap;
+                try {
+                    InputStream inputStream = HttpUtils.create(path).getInputStream();
+                    mBitmap = MyImageUtls.getSmallBitmap(inputStream, widthPixels, heightPixels);
+                } catch (IOException e) {
+                    e.printStackTrace();
+                    mBitmap = BitmapFactory.decodeResource(getActivity().getResources(),
+                            R.drawable.ic_pic_placeholder);
+                }
+                return mBitmap;
+            } else {
+                Bitmap mBitmap;
+                try {
+                    //                    long memoSize = MyImageUtls.getMemoSize(imgPath, widthPixels, heightPixels);
+                    mBitmap = MyImageUtls.getSmallBitmap(imgPath, widthPixels, heightPixels);
+                    BitmapFactory.decodeResource(getActivity().getResources(),
+                            R.drawable.ic_pic_placeholder);
+                } catch (Throwable e) {
+                    e.printStackTrace();
+                    mBitmap = BitmapFactory.decodeResource(getActivity().getResources(),
+                            R.drawable.ic_pic_placeholder);
+                }
+                if (mBitmap != null) {
+                    Log.e("zjy",
+                            tag + ": BitampSize==" + mBitmap.getByteCount() / 1024f / 1024);
+                }
+                return mBitmap;
+            }
+        }
     }
 }
