@@ -31,7 +31,7 @@ public class CheckUtils {
      @return true是，false否
      */
     public static boolean isAdmin() {
-        return ("101".equals(MyApp.id));
-//        return false;
+//        return ("101".equals(MyApp.id));
+        return false;
     }
 }
